requirements.txt
Django
djangorestframework
mysqlclient
django-mysql
psycopg2>=2.8
djangorestframework_simplejwt
django-seed
django-filter
weasyprint
celery
drf_spectacular
requests
watchdog